+++ components/BorrowModal/index.js
@@ -3,9 +3,9 @@ import { formatUnits } from "@ethersproject/units";
 import Tooltip from "@reach/tooltip";
 import { useWeb3React } from "@web3-react/core";
 import LENDING_MARKET_ABI from "constants/abis/lendingMarket.json";
-import useBorrow from "hooks/payables/useBorrow";
 import useCurrentToken from "hooks/useCurrentToken";
 import useMarketRegistryContract from "hooks/useMarketRegistryContract";
+import useBorrow from "hooks/calls/useBorrow";
 import useToast, { FLAVORS } from "hooks/useToast";
 import PropTypes from "prop-types";
 import { useEffect } from "react";
